requirements.txt
fake-bpy-module-latest
pyyaml
psutil
colorama
msvcrt
select